core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uiauto-core"
version = "1.0.0"
description = "Core automation engine for UI testing (framework-agnostic)"
readme = "README.md"
requires-python = ">=3.7"
authors = [
    {name = "UIAuto Contributors"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pyyaml>=5.4",
    "jsonschema>=4.0",
]

[project.urls]
Homepage = "https://github.com/AlperHuseyn/cita-uiauto-engine"
Repository = "https://github.com/AlperHuseyn/cita-uiauto-engine"

[tool.setuptools.packages.find]
where = ["."]
include = ["uiauto_core*"]

[tool.setuptools.package-data]
uiauto_core = ["schemas/*.json"]
